Added "Type" as Entity Type
Within the Belongs_To field is "type" which displays the phone's Owner as either a Person or a Business. This change appends this as a new column.

TO DO: Add a placeholder for resultRow that are not Owners. Currently, the entityType fills in as the same for each result number of a phone.